feat: View 추가
제일 마지막 레이어인 View를 구현했다. View Resolve 기능은 View 내부에서 메서드로 구현하였다. 모든 컨트롤러의 뷰 관련 로직의 프레임워크 의존성을 낮췄다.

+++ happyhouse/src/main/java/com/ssafy/happyhouse/controller/place/PlaceDeleteController.java
@@ -14,27 +14,14 @@ public class PlaceDeleteController implements Controller {
     private final FavPlaceService favPlaceService = FavPlaceServiceImpl.getInstace();
 
     @Override
-    public void get(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-    }
-
-    @Override
-    public void post(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    public String post(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         int favId = Integer.parseInt(request.getParameter("favId"));
 
-        int result = favPlaceService.deleteFavPlace(favId);
-
-        String memberId = (String) request.getSession().getAttribute("loginId");
-
         // 로그인 한 상태가 아니면 비정상적인 접근!
-        if (memberId == null) {
-            response.setContentType("text/html;charset=utf-8");
-            response.getWriter().write("<script>alert('로그인하고 접근해주세요!'); location.href='/member/login';</script>");
-            return;
-        }
+        if (request.getSession().getAttribute("loginId") == null) return "redirect:/member/login";
 
-        if (result != 0) {
-            response.sendRedirect("/place/items");
-        }
+        favPlaceService.deleteFavPlace(favId);
+
+        return "redirect:/place/items";
     }
 }
